+++ src/payments/mod.rs
@@ -1,4 +1,3 @@
 pub mod common;
 pub mod errors;
-pub mod ethereum;
 pub mod monero;